requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.40.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mysql-connector==2.2.9
mysql-connector-python==8.0.33
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==3.20.3
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
seaborn==0.12.2
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
